% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state()}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer or character. State by number}

\item{year}{integer or character}
}
\value{
The result of this function is a graph of a state map with accidents plotted in.
}
\description{
This function plots a state map of the accidents in a certain year
}
\note{
conditions that may result in an error: This function does not seem to work for state.num = 2: all regions out of bounds error
invalid state number generates an invalid state number message (for example state.num = 3)
no accidents generates a no accidents to plot message
}
\examples{
\dontrun{
    fars_map_state(1, 2013)
    fars_map_state("1", "2013")
    }

}
