% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read()}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character. The file name of the file to be loaded}
}
\value{
returns a tbl_df with the data
}
\description{
This function checks if the file exists, stops with a message if it does not or loads the file if it does.
}
\note{
conditions that may result in an error: if file does not exist -> "file does not exist" message
Also generates a default Warning message because of a deprecated function:
\code{tbl_df()} is deprecated as of dplyr 1.0.0.
Please use \code{tibble::as_tibble()} instead.
}
\examples{
\dontrun{
    fars_read(filename)
    fars_read(filename = "file name")
    }
}
