% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read()}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of integer or character}
}
\value{
NULL if there is an error

a list with of tibbles each with 2 colums: MONTH and year
}
\description{
This function reads in the file for each year and selects MONTH and year and returns these as a list of tibbles
}
\note{
conditions that may result in an error: input of an invalid year returns a warning
}
\examples{
\dontrun{
    fars_read_years(c(2013, 2014))
    fars_read_years(c("2013", "2014"))
    }
}
