% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years()}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of int or character}
}
\value{
tbl_df with MONTH in the first column and # of accidents for each month in each year in the following columns.
}
\description{
This function reads the data for each year and summarizes it into month and year columns.
}
\examples{
\dontrun{
    fars_summarize(c(2013, 2014))
    fars_summarize(c("2013", "2014"))
    }

}
