% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename()}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer or character. The file name of the file to be loaded}
}
\value{
character. Returns a file name
}
\description{
This function takes the year and makes it an integer, then uses it to construct the file name
}
\examples{
\dontrun{
    make_filename(2013)
    make_filename("2013")
    }
}
